pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["google-labs-jules[bot] <161369871+google-labs-jules[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = {extras = ["standard"], version = ">=0.34.3,<0.35.0"}
redis = ">=6.2.0,<7.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
langchain = ">=0.3.25,<0.4.0"
langchain-openai = ">=0.3.21,<0.4.0"
langchain-google-genai = ">=2.1.5,<3.0.0"
pydantic-settings = "^2.9.1"
# Dependencies for vendored ari-py
requests = ">=2.0.0,<3.0.0" # For ari_py's model.py and swaggerpy
websocket-client = "==1.8.0" # From ari-0.1.3 installation (used by swaggerpy for websockets)
langchain-community = "^0.3.24"
sqlalchemy = "^2.0.41"
jinja2 = "^3.1.6"
aiofiles = "^24.1.0"
python-multipart = "^0.0.20"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-cov = "^6.1.1"
httpx = "^0.28.1"
pytest-httpx = "^0.35.0"

[build-system]
requires = ["poetry-core>=1.0.0"] # Adjusted to a more common version for poetry-core
build-backend = "poetry.core.masonry.api"

# Remove the old [project] table as its relevant parts are moved to [tool.poetry]
# Remove [project.dependencies] as it's now [tool.poetry.dependencies]
